Parameter type

+++ app/src/main/java/one/mixin/android/tip/wc/WalletConnectV2.kt
@@ -474,7 +474,7 @@ object WalletConnectV2 : WalletConnect() {
             )
         } else {
             RawTransaction.createTransaction(
-                chain.chainReference.toLong(),
+                BigInteger.valueOf(chain.chainReference.toLong()),
                 nonce,
                 gasLimit,
                 transaction.to,
